requirements.txt
streamlit
torch
faiss-cpu
rank-bm25
sentence-transformers
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
langchain-huggingface
langchain-huggingface
pymupdf
python-dotenv
numpy
tqdm
datasets
